+++ lib/db/src/schema/index.ts
@@ -78,6 +78,7 @@ export * from "./compliance";
 export * from "./declaraties";
 export * from "./algemene-inkoop";
 export * from "./actiepunten";
+export * from "./acceptatieregister";
 export * from "./kalender";
 export * from "./voorkeuren";
 export * from "./werving";